dfc: Add a sequence id per client to skip missing transactions
When a client sends a requests to a subset of bricks, the transaction
id is incremented, but some bricks do not receive it. This produces
timeouts of future transactions.

To solve this a sequence number per client is used to keep track of
all requests sent to a specific brick.